+++ draft-packages/button/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@2.0.6...@kaizen/draft-button@2.0.7) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-button
+
+
+
+
+
+## [2.0.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@2.0.5...@kaizen/draft-button@2.0.6) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-button
+
+
+
+
+
 ## [2.0.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@2.0.4...@kaizen/draft-button@2.0.5) (2020-09-30)
 
 **Note:** Version bump only for package @kaizen/draft-button