+++ draft-packages/dropdown-menu/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.40](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.5.39...@kaizen/draft-dropdown-menu@1.5.40) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.5.39](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.5.38...@kaizen/draft-dropdown-menu@1.5.39) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
 ## [1.5.38](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.5.37...@kaizen/draft-dropdown-menu@1.5.38) (2020-09-30)
 
 **Note:** Version bump only for package @kaizen/draft-dropdown-menu