+++ draft-packages/hierarchical-select/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.23](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-select@1.3.22...@kaizen/draft-hierarchical-select@1.3.23) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-hierarchical-select
+
+
+
+
+
+## [1.3.22](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-select@1.3.21...@kaizen/draft-hierarchical-select@1.3.22) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-hierarchical-select
+
+
+
+
+
 ## [1.3.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-select@1.3.20...@kaizen/draft-hierarchical-select@1.3.21) (2020-09-30)
 
 **Note:** Version bump only for package @kaizen/draft-hierarchical-select