+++ legacy-packages/title-block/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.46](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@1.5.45...@kaizen/draft-title-block@1.5.46) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
+## [1.5.45](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@1.5.44...@kaizen/draft-title-block@1.5.45) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
 ## [1.5.44](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@1.5.43...@kaizen/draft-title-block@1.5.44) (2020-09-30)
 
 **Note:** Version bump only for package @kaizen/draft-title-block